Move getUrl from config to utils
