+++ src/config.test.js
@@ -1,28 +1,6 @@
-import { getUrl, localStorageName, flags } from './config';
+import { localStorageName, flags } from './config';
 
 describe('config', () => {
-  describe('getUrl', () => {
-    it('should throw an error when passed an invalid argument', () => {
-      expect(() => getUrl()).toThrow();
-      expect(() => getUrl('unknown')).toThrow();
-      expect(() => getUrl(null)).toThrow();
-    });
-
-    it('should return the "main" json file', () => {
-      expect(getUrl('main')).toEqual('./api/main');
-    });
-
-    it('should return a "pipeline" json file', () => {
-      const id = '123456';
-      expect(getUrl('pipeline', id)).toEqual(`./api/pipelines/${id}`);
-    });
-
-    it('should always return a relative path', () => {
-      expect(getUrl('main').substr(0, 2)).toEqual('./');
-      expect(getUrl('pipeline', 123).substr(0, 2)).toEqual('./');
-    });
-  });
-
   describe('localStorageName', () => {
     it('should contain KedroViz', () => {
       expect(localStorageName).toEqual(expect.stringContaining('KedroViz'));
