+++ src/utils/index.js
@@ -1,5 +1,7 @@
 //--- Useful JS utility functions ---//
 
+import { pathRoot } from '../config';
+
 /**
  * Loop through an array and output to an object
  * @param {Array} array
@@ -11,6 +13,23 @@ export const arrayToObject = (array, callback) => {
     newObject[key] = callback(key);
   });
   return newObject;
+};
+
+/**
+ * Determine the endpoint URL for loading different data types
+ * @param {string} type Data type
+ * @param {string=} id Endpoint identifier e.g. pipeline ID
+ */
+export const getUrl = (type, id) => {
+  switch (type) {
+    case 'main':
+      return [pathRoot, 'main'].join('/');
+    case 'pipeline':
+      if (!id) throw new Error('No pipeline ID provided');
+      return [pathRoot, 'pipelines', id].join('/');
+    default:
+      throw new Error('Unknown URL type');
+  }
 };
 
 /**
